Split printf.h according to 'poison pattern'
The pattern is to have two copies of an include file, in this case
printf.h and _printf.h. The latter is #included only in the
implementation file, and provides prototypes. The former #includes the
latter, but also poisons all the identifiers.

This pattern is because otherwise it's very easy to forget to use the e*
versions of functions. Using e.g. printf instead of eprintf is an easy
mistake to make, and will result in errors being silently ignored.